adding progressive and bye player
