+++ swiss_tournament/data/player.py
@@ -7,3 +7,7 @@ class Player:
 
     def __hash__(self):
         return hash(self.name)
+
+
+# Fictional player to pair with non-paired people in each round
+BYE = Player("bye")
