+++ swiss_tournament/step/helper_mappings.py
@@ -1,6 +1,7 @@
 from typing import List, Dict
 
 from generate import Player
+from swiss_tournament.data.player import BYE
 from swiss_tournament.data.tournament import Tournament
 
 
@@ -8,14 +9,18 @@ def player_round_map(tournament: Tournament) -> Dict[Player, List[tuple[Player, 
     mapping: Dict[Player, List[tuple[Player, float]]] = {}
     for player in tournament.players:
         mapping[player] = []
+    mapping[BYE] = []
     for rnd in tournament.rounds:
         for result in rnd:
             white = mapping[result.white]
             white.append((result.black, result.result))
             mapping[result.white] = white
             black = mapping[result.black]
-            black.append((result.white, result.result))
+            black.append((result.white, 1 - result.result))
             mapping[result.black] = black
+
+    # Remove results from fictional BYE player
+    mapping[BYE] = []
     return mapping
 
 
@@ -23,8 +28,12 @@ def player_points_map(tournament: Tournament) -> Dict[Player, float]:
     mapping: Dict[Player, float] = {}
     for player in tournament.players:
         mapping[player] = 0
+    mapping[BYE] = 0
     for rnd in tournament.rounds:
         for result in rnd:
             mapping[result.white] = mapping[result.white] + result.result
             mapping[result.black] = mapping[result.black] + 1 - result.result
+
+    # Remove results from fictional BYE player
+    mapping[BYE] = 0
     return mapping
