+++ tests/tiebreaks/test_progressive.py
@@ -1,13 +1,13 @@
 import unittest
 
-from swiss_tournament.data.player import Player, BYE
+from swiss_tournament.data.player import Player
 from swiss_tournament.data.result import Result
 from swiss_tournament.data.tournament import Tournament
-from swiss_tournament.step.tie_breaker import MedianBuchholz
+from swiss_tournament.step.tie_breaker import Progressive
 
 
-class MedianBuchholzTestCase(unittest.TestCase):
-    unit = MedianBuchholz()
+class ProgressiveTestCase(unittest.TestCase):
+    unit = Progressive()
 
     def test_without_rounds(self):
         result = self.unit.get(Player("alice"), Tournament([Player("alice")], []))
@@ -52,7 +52,7 @@ class MedianBuchholzTestCase(unittest.TestCase):
                 ]
             )
         )
-        self.assertEqual(0, result)
+        self.assertEqual(0.5, result)
 
     def test_with_three_rounds(self):
         result = self.unit.get(
@@ -66,7 +66,7 @@ class MedianBuchholzTestCase(unittest.TestCase):
                 ],
                 rounds=[
                     [
-                        Result(Player("alice"), Player("bob"), 1),
+                        Result(Player("alice"), Player("bob"), 0),
                         Result(Player("charlie"), Player("dave"), 1)
                     ],
                     [
@@ -80,60 +80,4 @@ class MedianBuchholzTestCase(unittest.TestCase):
                 ]
             )
         )
-        self.assertEqual(2, result)
-
-    def test_with_bye_on_player(self):
-        result = self.unit.get(
-            player=Player("bob"),
-            tournament=Tournament(
-                players=[
-                    Player("alice"),
-                    Player("bob"),
-                    Player("charlie"),
-                    Player("dave")
-                ],
-                rounds=[
-                    [
-                        Result(Player("alice"), Player("bob"), 1),
-                        Result(Player("charlie"), Player("dave"), 0.5)
-                    ],
-                    [
-                        Result(Player("charlie"), Player("alice"), 1),
-                        Result(Player("bob"), BYE, 0.5)
-                    ],
-                    [
-                        Result(Player("alice"), Player("dave"), 1),
-                        Result(Player("bob"), Player("dave"), 0.5)
-                    ]
-                ]
-            )
-        )
-        self.assertEqual(1, result)
-
-    def test_with_bye_on_opponent(self):
-        result = self.unit.get(
-            player=Player("bob"),
-            tournament=Tournament(
-                players=[
-                    Player("alice"),
-                    Player("bob"),
-                    Player("charlie"),
-                    Player("dave")
-                ],
-                rounds=[
-                    [
-                        Result(Player("alice"), Player("bob"), 1),
-                        Result(Player("charlie"), Player("dave"), 0.5)
-                    ],
-                    [
-                        Result(Player("alice"), BYE, 1),
-                        Result(Player("bob"), Player("charlie"), 0.5)
-                    ],
-                    [
-                        Result(Player("alice"), Player("charlie"), 1),
-                        Result(Player("bob"), Player("dave"), 0.5)
-                    ]
-                ]
-            )
-        )
-        self.assertEqual(1, result)
+        self.assertEqual(4.5, result)
